==> Python_Madness_2023/notebooks/step03_aScrapeBR.ipynb <==
# Scrape Basketball Reference 2008 to 2022
# https://www.sports-reference.com/cbb/seasons/2008-ratings.html
# Importing the required libraries
import requests
import pandas as pd
from bs4 import BeautifulSoup
import numpy as np

season = np.arange(2008,2023,1)

allBR = pd.DataFrame(columns = ['Year','Team','W','L','Pts','Opp','MOV','SOS','OSRS','DSRS','SRS'])

for y in season:
    # Downloading contents of the web page
    url = f"https://www.sports-reference.com/cbb/seasons/{y}-ratings.html"
    data = requests.get(url).text

    # Creating BeautifulSoup object
    soup = BeautifulSoup(data, 'html.parser')
    table = soup.find('table', class_='stats_table')

    # Collecting Ddata
    for row in table.tbody.find_all('tr'):    
        # Find all data for each column
        columns = row.find_all('td')
        if(columns != []):
            Team = columns[0].text.strip()
            W = columns[3].text.strip()
            L = columns[4].text.strip()
            Pts = columns[5].text.strip()
            Opp = columns[6].text.strip()
            MOV = columns[7].text.strip()
            SOS = columns[9].text.strip()
            OSRS = columns[11].text.strip()
            DSRS = columns[12].text.strip()
            SRS = columns[13].text.strip()
            newrow = [y,Team,W,L,Pts,Opp,MOV,SOS,OSRS,DSRS,SRS]
            allBR.loc[len(allBR)]=newrow
allBR.to_csv('step03_br0823.csv',index=False)

# Scrape Basketball Reference 2023
# https://www.sports-reference.com/cbb/seasons/2008-ratings.html
# Importing the required libraries
import requests
import pandas as pd
from bs4 import BeautifulSoup
import numpy as np

y = 2023

allBR = pd.read_csv('step03_br0823.csv')
allBR =allBR[allBR['Year']!=2023]


# Downloading contents of the web page
url = f"https://www.sports-reference.com/cbb/seasons/{y}-ratings.html"
data = requests.get(url).text

# Creating BeautifulSoup object
soup = BeautifulSoup(data, 'html.parser')
table = soup.find('table', class_='stats_table')

# Collecting Ddata
for row in table.tbody.find_all('tr'):    
    # Find all data for each column
    columns = row.find_all('td')
    if(columns != []):
        Team = columns[0].text.strip()
        W = columns[4].text.strip()
        L = columns[5].text.strip()
        Pts = columns[6].text.strip()
        Opp = columns[7].text.strip()
        MOV = columns[8].text.strip()
        SOS = columns[10].text.strip()
        OSRS = columns[12].text.strip()
        DSRS = columns[13].text.strip()
        SRS = columns[14].text.strip()
        newrow = [y,Team,W,L,Pts,Opp,MOV,SOS,OSRS,DSRS,SRS]
        allBR.loc[len(allBR)]=newrow
allBR.to_csv('step03_br0823.csv',index=False)
